+++ crates/runner-nordic/src/board/platform.rs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 use alloc::borrow::Cow;
+use alloc::string::String;
 
 use header::Header;
 use nrf52840_hal::pac::FICR;
@@ -21,6 +22,8 @@ use wasefire_board_api::platform::Api;
 use wasefire_common::addr_of_symbol;
 use wasefire_common::platform::Side;
 use wasefire_error::Code;
+use wasefire_logger as log;
+use wasefire_protocol::common::Name;
 use wasefire_protocol::platform::SideInfo;
 use wasefire_sync::Once;
 
@@ -44,7 +47,7 @@ impl Api for Impl {
     fn running_info() -> SideInfo<'static> {
         let side = Self::running_side();
         let header = Header::new(side);
-        let name = header.name().to_vec().into();
+        let name = build_name(header.name());
         let version = header.version().to_be_bytes().to_vec().into();
         SideInfo { name, version }
     }
@@ -58,7 +61,7 @@ impl Api for Impl {
         if !header.has_firmware() {
             return Err(Error::world(Code::NotFound));
         }
-        let name = header.name().to_vec().into();
+        let name = build_name(header.name());
         let version = header.version().to_be_bytes().to_vec().into();
         Ok(SideInfo { name, version })
     }
@@ -68,7 +71,20 @@ impl Api for Impl {
     }
 }
 
-pub fn reboot() -> ! {
+fn build_name(input: [u8; 24]) -> Name<'static> {
+    let output = try {
+        let len = input.iter().position(|&x| x == 0).unwrap_or(input.len());
+        input[len ..].iter().all(|&x| x == 0).then_some(())?;
+        Name::new(String::from_utf8(input[.. len].to_vec()).ok()?.into()).ok()?
+    };
+    if let Some(name) = output {
+        return name;
+    }
+    log::error!("Invalid name: {=[u8]:02x}", input);
+    Name::default()
+}
+
+fn reboot() -> ! {
     wasefire_logger::flush();
     nrf52840_hal::pac::SCB::sys_reset()
 }
@@ -82,7 +98,7 @@ pub fn init_serial(ficr: &FICR) {
     SERIAL.call_once(|| serial);
 }
 
-pub fn serial() -> &'static [u8; 8] {
+fn serial() -> &'static [u8; 8] {
     SERIAL.get().unwrap()
 }